move header and footer to sections/
