+++ src/components/page/page.js
@@ -1,8 +1,8 @@
 import React from "react"
 import { Helmet } from "react-helmet"
 import { useStaticQuery, graphql } from "gatsby"
-import Header from "../header"
-import Footer from "../footer"
+import Header from "../../sections/header"
+import Footer from "../../sections/footer"
 import Spacer from "../spacer"
 
 // global style
